server/requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.7
django-cors-headers==4.4.0
django-import-export==4.1.1
djangorestframework==3.15.2
loguru==0.7.2
pillow==10.4.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
